Remove duplicate stylesheet by setting the id
We were checking for the existence of an element with id
'braintree-dropin-stylesheet' but we failed to actually set the id
anywhere.

The test app was including both a local stylesheet and one from the
assets server. This fixes the issue by not including the assets server
stylesheet when the local one is already present.