Quick fix for your consideration, fixes #474, addresses #448
Here are a few simple changes that fix the weird styling of the "disabled" payment method in the Vault Manager and the "undefined" tooltip text issue. Also included is a small change that prevents the payment method associated with a subscription from being removed.

I think of this as a quick fix. If more options for the Vault Manager are planned for the future, I feel these changes address the reported problems adequately in the meantime.

Feel free to take what you want, leave what you don't, and I am also free to make further minor improvements if requested.