Replace placeholders with strings in dist
The JSDoc plugin was replacing placeholders in the documentation, but
ultimately that left placeholders all over the dist code outside of the
docs. This change interpolates the current version numbers not just in
the docs, but also all of the comments that the docs are generated
from.

Also apparently I accidentally wrote "statistics" instead of "statics"
in one place on the gulp@4 upgrade.